+++ docker-compose.yml
@@ -24,8 +24,8 @@ services:
       - "3000:3000"
     volumes:
       - ./frontend:/app
-    # environment:
-      # - CHOKIDAR_USEPOLLING=true
+    environment:
+      - CHOKIDAR_USEPOLLING=true
       # - WATCHPACK_POLLING=true
       
   db:
