+++ frontend/src/pages/Dashboard/Dashboard.tsx
@@ -1,0 +1,42 @@
+
+export default function Dashboard() {
+  return (
+    <a
+      href="https://tailadmin.com/pricing"
+      target="_blank"
+      rel="nofollow"
+      className="flex items-center justify-center p-3 font-medium text-white rounded-lg bg-brand-500 text-theme-sm hover:bg-brand-600"
+    >
+      Purchase Plan
+    </a>
+  );
+
+}
+
+// export default function Dashboard() {
+//   // const { tasks, activityData } = useDashboardViewModel();
+
+//   return (
+//     <div className="grid grid-cols-[250px_1fr] min-h-screen">
+//       <aside className="bg-gray-100 p-4 text-sm border-r">
+//         <div className="font-bold mb-4">Menú</div>
+//         <ul className="space-y-2">
+//           <li>🏠 Inicio</li>
+//           <li>📝 Mis taryaeas</li>
+//           <li>📊 Monitoreo</li>
+//         </ul>
+
+//         <div className="mt-6 font-bold">TeamSpace ➕</div>
+//         <ul className="ml-2 space-y-1">
+//           <li>📁 School</li>
+//           <li>📁 Taskfy</li>
+//         </ul>
+
+//         <div className="mt-6 font-bold">Otros</div>
+//         <ul className="ml-2">
+//           <li>⚙️ Settings</li>
+//         </ul>
+//       </aside>
+//     </div>
+//   );
+// };
